+++ src/packet/bwe/mod.rs
@@ -6,7 +6,6 @@
 
 use std::cmp::Ordering;
 use std::fmt;
-use std::ops::{Deref, RangeInclusive};
 use std::time::{Duration, Instant};
 
 use crate::rtp_::{Bitrate, DataSize, SeqNo, TwccSendRecord};
@@ -39,7 +38,6 @@ pub struct SendSideBandwithEstimator {
     loss_controller: Option<LossController>,
     acked_bitrate_estimator: AckedBitrateEstimator,
     started_at: Option<Instant>,
-    acked_packets_deduper: HandledPacketsTracker,
 }
 
 impl SendSideBandwithEstimator {
@@ -57,7 +55,6 @@ impl SendSideBandwithEstimator {
                 BITRATE_WINDOW,
             ),
             started_at: None,
-            acked_packets_deduper: HandledPacketsTracker::default(),
         }
     }
 
@@ -69,12 +66,7 @@ impl SendSideBandwithEstimator {
     ) {
         let _ = self.started_at.get_or_insert(now);
 
-        let send_records: Vec<_> = records
-            .filter(|r| {
-                // Skip acked packets that have already been processed before.
-                !self.acked_packets_deduper.contains(r.seq())
-            })
-            .collect();
+        let send_records: Vec<_> = records.collect();
         let mut acked_packets = vec![];
 
         let mut max_rtt = None;
@@ -82,12 +74,11 @@ impl SendSideBandwithEstimator {
         let mut lost = 0;
         for record in send_records.iter() {
             count += 1;
-            let Ok(acked_packet) = AckedPacket::try_from(*record) else {
+            let Ok(acked_packet) = (*record).try_into() else {
                 lost += 1;
                 continue;
             };
             acked_packets.push(acked_packet);
-            self.acked_packets_deduper.add(acked_packet.seq_no);
             max_rtt = max_rtt.max(record.rtt());
         }
         acked_packets.sort_by(AckedPacket::order_by_receive_time);
@@ -237,71 +228,6 @@ impl fmt::Display for BandwidthUsage {
             BandwidthUsage::Overuse => write!(f, "overuse"),
             BandwidthUsage::Normal => write!(f, "normal"),
             BandwidthUsage::Underuse => write!(f, "underuse"),
-        }
-    }
-}
-
-/// Sliding window [`SeqNo`]s tracker.
-struct HandledPacketsTracker {
-    /// Range of currently tracked [`SeqNo`]s.
-    window: RangeInclusive<SeqNo>,
-
-    /// Bit array of recently added packets.
-    history: [u8; 64],
-}
-
-impl HandledPacketsTracker {
-    /// Remembers the give [`SeqNo`].
-    ///
-    /// Expects somewhat sequential data since window always advances to hold
-    /// latest added value forgetting older ones.
-    pub fn add(&mut self, seq: SeqNo) {
-        self.maybe_advance_window(seq);
-
-        let (byte_idx, bit_idx) = self.pos_of_seq(seq);
-        self.history[byte_idx] |= 1 << bit_idx;
-    }
-
-    /// Checks if the provided [`SeqNo`] has been seen in the window.
-    pub fn contains(&self, seq: SeqNo) -> bool {
-        if self.window.contains(&seq) {
-            let (byte_idx, bit_idx) = self.pos_of_seq(seq);
-            (self.history[byte_idx] & (1 << bit_idx)) != 0
-        } else {
-            false
-        }
-    }
-
-    /// Advances the window to include the given [`SeqNo`].
-    fn maybe_advance_window(&mut self, new_max_seq: SeqNo) {
-        if new_max_seq <= *self.window.end() {
-            return;
-        }
-        // Clear newly included bits
-        for i in **self.window.end() + 1..*new_max_seq {
-            let (byte_idx, bit_idx) = self.pos_of_seq(&i);
-            self.history[byte_idx] &= !(1 << bit_idx);
-        }
-        let new_start = new_max_seq.saturating_sub(self.history.len() as u64 * 8);
-        self.window = RangeInclusive::new(SeqNo::from(new_start), new_max_seq);
-    }
-
-    /// Maps a given sequence number to its position in the bit vector.
-    fn pos_of_seq(&self, seq: impl Deref<Target = u64>) -> (usize, u8) {
-        let byte_idx = (*seq / 8) as usize % self.history.len();
-        let bit_idx = (*seq % 8) as u8;
-
-        (byte_idx, bit_idx)
-    }
-}
-
-impl Default for HandledPacketsTracker {
-    fn default() -> Self {
-        let history = [0; 64];
-
-        Self {
-            window: RangeInclusive::new(SeqNo::from(0), SeqNo::from(history.len() as u64 * 8)),
-            history,
         }
     }
 }