+++ CHANGELOG.md
@@ -22,6 +22,7 @@
   * Do not disconnect whilst we still check new candidates #489
   * Ensure lexical ordering of SDP-formatted candidates follows priority #557
   * Limit TWCC iteration with packet status count #606
+  * Dedupe acked packets from `TwccSendRegister::apply_report()` #601, #605
 
 # 0.6.2